requirements.txt
annotated-types   == 0.7.0
anyio             == 4.12.1
certifi           == 2026.2.25
colorama          == 0.4.6
distro            == 1.9.0
fsspec            == 2024.12.0
h11               == 0.16.0
httpcore          == 1.0.9
httpx             == 0.28.1
idna              == 3.11
jiter             == 0.13.0
json_repair       == 0.44.1
neo4j             == 6.1.0
neo4j-graphrag    == 1.14.0
numpy             == 2.4.3
openai            == 1.109.1
pip               == 26.0.1
pydantic          == 2.12.5
pydantic_core     == 2.41.5
pypdf             == 6.9.1
python-dotenv     == 1.2.2
pytz              == 2026.1.post1
PyYAML            == 6.0.3
scipy             == 1.17.1
sniffio           == 1.3.1
tenacity          == 9.1.4
tqdm              == 4.67.3
types-PyYAML      == 6.0.12.20250915
typing_extensions == 4.15.0
typing-inspection == 0.4.2